Monthly backup fails on first of month
The logic was attempting to create an archive for the previous
month using the archive from the first day of the current month.
As this doesn't exist till the beginning of the second day of the
month, the monthly archive was failing on the first, but succeeding
on the second.

Fix the logic so that the archive from the last day in the month
is used to create that months backup.

Signed-off-by: Martyn Welch <martyn@welchs.me.uk>